Fix nested Any() not correctly tracking longestError. Fixes #3
When nested the inner Any() can end up seeing Error.pos set, but Error.expecting to be "". This gets copied into longestError.  This can subsequently lead it to not update longestError after calling a child parser (because the child error is before longestError). This results in it incorrectly indicating there was no error.